% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_reduction_dynamique.R
\name{compute_dynamic_energy}
\alias{compute_dynamic_energy}
\title{Calcul de l'energie dynamique d'une image}
\usage{
compute_dynamic_energy(gray_img)
}
\arguments{
\item{gray_img}{Image en niveaux de gris (cimg)}
}
\value{
Image d'energie (cimg)
}
\description{
Calcul de l'energie dynamique d'une image
}
