% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_functions.R, R/image_reduction.R
\name{compute_energy}
\alias{compute_energy}
\title{Compute energy matrix}
\usage{
compute_energy(gray_img)
}
\arguments{
\item{gray_img}{Image en niveaux de gris (cimg)}
}
\value{
Energy matrix

Image d'energie
}
\description{
C++ implementation of energy calculation
}
