% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_reduction_dynamique.R
\name{cumulative_dynamic_energy}
\alias{cumulative_dynamic_energy}
\title{Calcule la matrice cumulative d'energie dynamique}
\usage{
cumulative_dynamic_energy(energy_mat)
}
\arguments{
\item{energy_mat}{Matrice d'energie (h x w)}
}
\value{
Matrice cumulative (h x w)
}
\description{
Calcule la matrice cumulative d'energie dynamique
}
