% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_reduction_dynamique.R
\name{find_dynamic_seam}
\alias{find_dynamic_seam}
\title{Retrouve la seam minimale dynamique}
\usage{
find_dynamic_seam(cum_energy)
}
\arguments{
\item{cum_energy}{Matrice cumulative d'energie (h x w)}
}
\value{
Vecteur (longueur h) representant la seam
}
\description{
Retrouve la seam minimale dynamique
}
