% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_reduction_dynamique.R
\name{reduce_dynamic_image_seams}
\alias{reduce_dynamic_image_seams}
\title{Reduction d'image par seam carving dynamique}
\usage{
reduce_dynamic_image_seams(im, n_seams)
}
\arguments{
\item{im}{Image a reduire (objet `cimg`)}

\item{n_seams}{Nombre de coutures a supprimer}
}
\value{
Image reduite (objet `cimg`)
}
\description{
Reduction d'image par seam carving dynamique
}
