% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_functions.R, R/image_reduction.R
\name{reduce_image_seams}
\alias{reduce_image_seams}
\title{Reduce image by seams}
\usage{
reduce_image_seams(im, n_seams = 10)
}
\arguments{
\item{im}{Image (objet `cimg`)}

\item{n_seams}{Nombre de coutures a supprimer}
}
\value{
Reduced image

Image reduite
}
\description{
C++ implementation of seam carving
}
