% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_reduction_dynamique.R
\name{remove_dynamic_vertical_seam}
\alias{remove_dynamic_vertical_seam}
\title{Supprime une couture verticale dynamique}
\usage{
remove_dynamic_vertical_seam(im, seam)
}
\arguments{
\item{im}{Image (objet `cimg`)}

\item{seam}{Vecteur des indices de colonnes a supprimer}
}
\value{
Image modifiee (objet `cimg`)
}
\description{
Supprime une couture verticale dynamique
}
