% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_reduction.R
\name{seam_carving}
\alias{seam_carving}
\title{Applique le seam carving}
\usage{
seam_carving(im, n_seams = 10)
}
\arguments{
\item{im}{Image (objet `cimg`)}

\item{n_seams}{Nombre de coutures a supprimer}
}
\value{
Image reduite
}
\description{
Applique le seam carving
}
