% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_reduction.R
\name{trace_greedy_seam}
\alias{trace_greedy_seam}
\title{Trace une couture gloutonne}
\usage{
trace_greedy_seam(energy_mat, start_col, penalty = 0)
}
\arguments{
\item{energy_mat}{Matrice d'energie (h x w)}

\item{start_col}{Colonne de depart}

\item{penalty}{Penalite pour diagonales}
}
\value{
Vecteur d'indices de colonnes
}
\description{
Trace une couture gloutonne
}
